requirements.txt
fild==0.0.9
fild-api-hub==0.0.4
pytest~=8.3.3
